Create python code
It is a mini-game. Make a program in which the computer randomly chooses a number between 1 to 100. Then give users a hint to guess the number. Every time the user guesses wrong, he gets another clue, and his score gets reduced. The clue can be multiples, divisible, greater or smaller, or a combination of all.